@@ -1,270 +1,309 @@
 {
-  "name": "postgres.garden",
+  "name": "postgres-garden",
   "private": true,
   "version": "0.0.0",
-  "author": "Daniel Gray<danielfgray@gmail.com>",
-  "License": "MIT",
   "type": "module",
+  "projectName": "Postgres Garden",
+  "fromEmail": "noreply@postgres.garden",
+  "legalText": "Postgres Garden - Open Source PostgreSQL Development Environment",
   "scripts": {
-    "dev:client": "bun -b vite --clearScreen false",
-    "dev:server": "bun --hot --no-clear-screen server/index.ts",
-    "predev": "bun ./scripts/dbStart.mjs",
-    "dev": "run-p dev:client dev:server db:watch worker",
-    "test:watch": "bun test --watch --no-clear-screen",
-    "test:units": "bun test",
-    "test:e2e": "cypress run --headless",
-    "test:browser": "cypress open --e2e",
-    "test": "run-s test:units",
-    "worker": "cd worker && bun -b --env-file=../.env graphile-worker",
-    "fmt": "prettier --write .",
-    "lint:js": "eslint --ignore-path .gitignore --cache-file node_modules/.cache/eslintcache --ext .js,.ts,.mjs .",
-    "lint": "run-s lint:js",
-    "typecheck:watch": "tsc --noEmit --pretty --watch --preserveWatchOutput",
-    "build:tsc": "tsc",
-    "build:server": "esbuild server/index.ts --bundle --platform=node --format=esm --outfile=dist/server.js --sourcemap",
+    "dev:client": "vite dev",
+    "dev": "mprocs 'bun dev:server' 'bun db:watch' 'bun worker' 'bun test:watch' 'bun typecheck:watch' 'bun cypress:browser' 'bun build:renderer --watch' 'bun build:webviews --watch'",
+    "dev:server": "bun --hot server/index.ts 2>&1 | tee -a ./server.log",
+    "predev": "run-s db:ensure",
+    "prebuid": "rm -fr dist",
+    "build:server": "NODE_ENV=production bun scripts/serverBuild.ts",
+    "build:renderer": "vite build --config vite.renderer.config.ts",
+    "build:webviews": "vite build --config vite.webviews.config.ts",
     "build:client": "vite build",
-    "build": "run-s build:client",
-    "server:start": "bun server/index.ts",
-    "start": "run-s server:start worker",
-    "preview": "vite preview",
-    "env:init": "bun scripts/generateEnvFile.mjs",
+    "build": "run-s build:server build:webviews build:renderer build:client",
+    "start": "NODE_ENV=production bun run server/index.ts",
+    "migrate": "bun -b graphile-migrate migrate",
+    "start:prod": "bun run migrate && bun run start",
+    "init:prod": "NOCONFIRM=1 bun scripts/dbSetup.ts && bun -b graphile-migrate reset --erase",
+    "lint": "eslint .",
+    "lint:fix": "eslint . --fix",
+    "typecheck:watch": "tsc --noEmit --watch",
+    "typecheck": "tsc --noEmit",
+    "start:debugServer": "bun src/debugServer.ts",
+    "cypress:chromium": "cypress open --e2e --browser=chromium",
+    "cypress:firefox": "cypress open --e2e --browser=firefox",
+    "cypress:browser": "cypress open --e2e",
+    "cypress:headless": "cypress run",
+    "test:unit": "bun test server/",
+    "test:watch": "bun test --watch server/",
+    "test:e2e": "run-s cypress:headless",
+    "test": "run-s test:unit test:e2e",
+    "worker": "cd worker && bun --env-file=../.env graphile-worker",
+    "worker:schema": "cd worker && bun --env-file=../.env graphile-worker --schema-only",
+    "env:init": "bun scripts/generateEnvFile.ts",
     "gm": "bun -b graphile-migrate",
-    "db:init": "bun scripts/setup.mjs",
-    "db:start": "docker-compose up -d db",
+    "db:init": "bun scripts/dbSetup.ts",
+    "db:ensure": "bun scripts/dbEnsure.ts",
+    "db:up": "docker-compose up -d db",
     "db:latest": "bun -b graphile-migrate watch --once",
     "db:watch": "bun -b graphile-migrate watch",
     "db:reset": "bun -b graphile-migrate reset --erase",
     "db:stop": "docker-compose stop db",
     "db:exec": "docker-compose exec db",
-    "db:setup": "run-s db:start db:init db:reset db:latest",
-    "setup": "NOCONFIRM=1 run-s env:init db:setup"
+    "db:types": "kysely-codegen --dialect postgres --exclude-pattern 'graphile*.*' --out-file ./generated/db.ts",
+    "db:setup": "run-s db:up db:init db:reset db:latest",
+    "init": "NOCONFIRM=1 run-s env:init db:setup worker:schema"
   },
   "dependencies": {
-    "@codingame/monaco-vscode-accessibility-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-ai-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-all-default-extensions": "^16.0.6",
-    "@codingame/monaco-vscode-all-language-default-extensions": "^16.0.6",
-    "@codingame/monaco-vscode-all-language-feature-default-extensions": "^16.0.6",
-    "@codingame/monaco-vscode-api": "^16.0.6",
-    "@codingame/monaco-vscode-authentication-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-base-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-bat-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-bulk-edit-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-chat-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-clojure-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-coffeescript-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-comments-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-configuration-editing-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-configuration-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-cpp-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-csharp-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-css-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-css-language-features-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-dart-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-debug-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-dialogs-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-diff-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-docker-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-edit-sessions-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-editor-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-emmet-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-emmet-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-environment-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-explorer-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-extension-editing-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-extension-gallery-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-extensions-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-files-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-fsharp-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-git-base-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-github-authentication-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-go-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-groovy-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-handlebars-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-hlsl-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-host-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-html-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-html-language-features-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-ini-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-interactive-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-ipynb-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-issue-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-java-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-javascript-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-json-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-json-language-features-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-julia-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-keybindings-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-language-detection-worker-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-language-pack-cs": "^16.0.6",
-    "@codingame/monaco-vscode-language-pack-de": "^16.0.6",
-    "@codingame/monaco-vscode-language-pack-es": "^16.0.6",
-    "@codingame/monaco-vscode-language-pack-fr": "^16.0.6",
-    "@codingame/monaco-vscode-language-pack-it": "^16.0.6",
-    "@codingame/monaco-vscode-language-pack-ja": "^16.0.6",
-    "@codingame/monaco-vscode-language-pack-ko": "^16.0.6",
-    "@codingame/monaco-vscode-language-pack-pl": "^16.0.6",
-    "@codingame/monaco-vscode-language-pack-pt-br": "^16.0.6",
-    "@codingame/monaco-vscode-language-pack-qps-ploc": "^16.0.6",
-    "@codingame/monaco-vscode-language-pack-ru": "^16.0.6",
-    "@codingame/monaco-vscode-language-pack-tr": "^16.0.6",
-    "@codingame/monaco-vscode-language-pack-zh-hans": "^16.0.6",
-    "@codingame/monaco-vscode-language-pack-zh-hant": "^16.0.6",
-    "@codingame/monaco-vscode-languages-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-latex-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-layout-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-less-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-lifecycle-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-localization-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-log-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-log-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-lua-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-make-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-markdown-basics-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-markdown-language-features-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-markdown-math-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-markers-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-mcp-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-media-preview-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-merge-conflict-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-microsoft-authentication-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-model-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-monarch-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-multi-diff-editor-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-notebook-renderers-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-notebook-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-notifications-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-npm-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-objective-c-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-outline-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-output-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-performance-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-perl-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-php-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-powershell-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-preferences-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-pug-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-python-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-quickaccess-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-r-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-razor-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-references-view-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-relauncher-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-remote-agent-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-restructuredtext-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-ruby-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-rust-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-scm-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-scss-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-search-result-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-search-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-secret-storage-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-shaderlab-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-share-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-shellscript-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-simple-browser-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-snippets-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-speech-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-sql-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-standalone-css-language-features": "^16.0.6",
-    "@codingame/monaco-vscode-standalone-html-language-features": "^16.0.6",
-    "@codingame/monaco-vscode-standalone-json-language-features": "^16.0.6",
-    "@codingame/monaco-vscode-standalone-languages": "^16.0.6",
-    "@codingame/monaco-vscode-standalone-typescript-language-features": "^16.0.6",
-    "@codingame/monaco-vscode-storage-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-survey-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-swift-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-task-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-telemetry-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-terminal-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-testing-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-textmate-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-theme-abyss-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-theme-defaults-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-theme-kimbie-dark-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-theme-monokai-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-theme-monokai-dimmed-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-theme-quietlight-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-theme-red-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-theme-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-theme-seti-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-theme-solarized-dark-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-theme-solarized-light-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-theme-tomorrow-night-blue-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-timeline-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-treesitter-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-typescript-basics-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-typescript-language-features-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-update-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-user-data-profile-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-user-data-sync-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-vb-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-view-banner-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-view-common-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-view-status-bar-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-view-title-bar-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-views-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-walkthrough-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-welcome-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-workbench-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-working-copy-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-workspace-trust-service-override": "^16.0.6",
-    "@codingame/monaco-vscode-xml-default-extension": "^16.0.6",
-    "@codingame/monaco-vscode-yaml-default-extension": "^16.0.6",
-    "@electric-sql/pglite": "^0.3.0",
-    "@elysiajs/eden": "^1.2.0",
-    "@lucia-auth/adapter-postgresql": "^3.1.2",
-    "@octokit/webhooks": "^13.4.1",
-    "arctic": "^1.9.2",
-    "classnames": "^2.5.1",
-    "dockerode": "^4.0.2",
-    "dotenv": "^16.4.7",
-    "elysia": "^1.2.9",
+    "@aws-sdk/client-sesv2": "^3.990.0",
+    "@codingame/monaco-vscode-accessibility-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-ai-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-all-default-extensions": "^26.0.0",
+    "@codingame/monaco-vscode-all-language-default-extensions": "^26.0.0",
+    "@codingame/monaco-vscode-all-language-feature-default-extensions": "^26.0.0",
+    "@codingame/monaco-vscode-api": "^26.0.0",
+    "@codingame/monaco-vscode-assignment-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-authentication-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-base-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-bat-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-bulk-edit-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-chat-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-clojure-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-coffeescript-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-comments-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-configuration-editing-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-configuration-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-cpp-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-csharp-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-css-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-css-language-features-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-dart-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-debug-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-dialogs-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-diff-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-docker-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-edit-sessions-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-editor-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-emmet-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-emmet-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-environment-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-explorer-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-extension-editing-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-extension-gallery-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-extensions-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-files-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-fsharp-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-git-base-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-github-authentication-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-go-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-groovy-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-handlebars-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-hlsl-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-host-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-html-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-html-language-features-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-image-resize-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-ini-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-interactive-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-ipynb-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-issue-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-java-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-javascript-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-json-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-json-language-features-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-julia-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-keybindings-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-language-detection-worker-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-language-pack-cs": "^26.0.0",
+    "@codingame/monaco-vscode-language-pack-de": "^26.0.0",
+    "@codingame/monaco-vscode-language-pack-es": "^26.0.0",
+    "@codingame/monaco-vscode-language-pack-fr": "^26.0.0",
+    "@codingame/monaco-vscode-language-pack-it": "^26.0.0",
+    "@codingame/monaco-vscode-language-pack-ja": "^26.0.0",
+    "@codingame/monaco-vscode-language-pack-ko": "^26.0.0",
+    "@codingame/monaco-vscode-language-pack-pl": "^26.0.0",
+    "@codingame/monaco-vscode-language-pack-pt-br": "^26.0.0",
+    "@codingame/monaco-vscode-language-pack-qps-ploc": "^26.0.0",
+    "@codingame/monaco-vscode-language-pack-ru": "^26.0.0",
+    "@codingame/monaco-vscode-language-pack-tr": "^26.0.0",
+    "@codingame/monaco-vscode-language-pack-zh-hans": "^26.0.0",
+    "@codingame/monaco-vscode-language-pack-zh-hant": "^26.0.0",
+    "@codingame/monaco-vscode-languages-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-latex-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-layout-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-less-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-lifecycle-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-localization-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-log-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-log-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-lua-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-make-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-markdown-basics-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-markdown-language-features-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-markdown-math-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-markers-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-mcp-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-media-preview-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-merge-conflict-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-mermaid-chat-features-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-model-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-monarch-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-multi-diff-editor-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-notebook-renderers-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-notebook-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-notifications-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-npm-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-objective-c-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-outline-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-output-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-performance-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-perl-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-php-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-powershell-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-preferences-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-process-explorer-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-prompt-basics-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-pug-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-python-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-quickaccess-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-r-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-razor-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-references-view-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-relauncher-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-remote-agent-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-restructuredtext-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-ruby-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-rust-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-scm-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-scss-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-search-result-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-search-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-secret-storage-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-shaderlab-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-share-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-shellscript-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-simple-browser-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-snippets-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-speech-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-sql-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-standalone-css-language-features": "^26.0.0",
+    "@codingame/monaco-vscode-standalone-html-language-features": "^26.0.0",
+    "@codingame/monaco-vscode-standalone-json-language-features": "^26.0.0",
+    "@codingame/monaco-vscode-standalone-languages": "^26.0.0",
+    "@codingame/monaco-vscode-standalone-typescript-language-features": "^26.0.0",
+    "@codingame/monaco-vscode-storage-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-survey-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-swift-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-task-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-telemetry-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-terminal-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-testing-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-textmate-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-theme-2026-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-theme-abyss-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-theme-defaults-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-theme-kimbie-dark-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-theme-monokai-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-theme-monokai-dimmed-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-theme-quietlight-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-theme-red-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-theme-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-theme-seti-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-theme-solarized-dark-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-theme-solarized-light-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-theme-tomorrow-night-blue-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-timeline-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-treesitter-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-typescript-basics-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-typescript-language-features-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-update-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-user-data-profile-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-user-data-sync-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-vb-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-view-banner-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-view-common-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-view-status-bar-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-view-title-bar-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-views-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-walkthrough-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-welcome-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-workbench-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-working-copy-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-workspace-trust-service-override": "^26.0.0",
+    "@codingame/monaco-vscode-xml-default-extension": "^26.0.0",
+    "@codingame/monaco-vscode-yaml-default-extension": "^26.0.0",
+    "@effect/ai": "^0.33.2",
+    "@effect/ai-anthropic": "^0.23.0",
+    "@effect/ai-google": "^0.12.1",
+    "@effect/ai-openai": "^0.37.2",
+    "@effect/ai-openrouter": "^0.8.3",
+    "@effect/cli": "^0.73.2",
+    "@effect/platform": "^0.94.5",
+    "@effect/platform-browser": "^0.74.0",
+    "@effect/platform-node": "^0.104.1",
+    "@electric-sql/pglite": "^0.3.15",
+    "@elysiajs/cookie": "^0.8.0",
+    "@elysiajs/eden": "^1.4.8",
+    "@elysiajs/static": "^1.4.7",
+    "@octokit/webhooks": "^14.2.0",
+    "@postgres-language-server/wasm": "^0.21.0",
+    "@preact/signals": "^2.8.0",
+    "@types/vscode-notebook-renderer": "^1.72.4",
+    "@virtualstate/navigation": "^1.0.1-alpha.212",
+    "ansi-colors": "^4.1.3",
+    "arctic": "^3.7.0",
+    "bootstrap": "^5.3.8",
+    "dockerode": "^4.0.9",
+    "dotenv": "^16.6.1",
+    "effect": "^3.19.17",
+    "elysia": "^1.4.25",
     "graphile-migrate": "^2.0.0-rc.2",
     "graphile-worker": "^0.16.6",
     "html-to-text": "^9.0.5",
-    "logestic": "^1.2.4",
-    "lucia": "^3.2.2",
-    "mermaid": "^11.4.1",
-    "mjml": "^4.15.3",
-    "monaco-editor": "npm:@codingame/monaco-vscode-editor-api",
-    "nodemailer": "^6.9.16",
+    "kysely": "^0.28.11",
+    "kysely-codegen": "^0.19.0",
+    "lodash": "^4.17.23",
+    "lodash.template": "^4.5.0",
+    "mermaid": "^11.12.2",
+    "mjml": "^4.18.0",
+    "monaco-editor": "npm:@codingame/monaco-vscode-editor-api@^26.0.0",
+    "mprocs": "^0.8.3",
+    "nodemailer": "^7.0.12",
     "npm-run-all": "^4.1.5",
-    "pg-introspection": "^0.0.1-beta.10",
-    "postgres": "github:DanielFGray/postgres.js",
+    "pev2": "^1.20.1",
+    "pg": "^8.18.0",
+    "pg-introspection": "^1.0.0-rc.4",
     "postgres-semicolons": "^0.1.2",
-    "sql-formatter": "^15.3.2",
-    "sql-parser-cst": "^0.28.0",
-    "vscode": "npm:@codingame/monaco-vscode-extension-api",
-    "ws": "^8.18.0",
-    "zod": "^3.24.1"
+    "preact": "^10.28.3",
+    "svg-pan-zoom": "^3.6.2",
+    "urlpattern-polyfill": "^10.1.0",
+    "vscode": "npm:@codingame/monaco-vscode-extension-api@^26.0.0",
+    "vscode-notebook-error-overlay": "^1.1.0",
+    "vue": "^3.5.28",
+    "ws": "^8.19.0"
   },
   "devDependencies": {
     "@codingame/esbuild-import-meta-url-plugin": "^1.0.3",
-    "@codingame/monaco-vscode-rollup-extension-directory-plugin": "^16.0.6",
-    "@codingame/monaco-vscode-rollup-vsix-plugin": "^16.0.6",
-    "@types/bun": "^1.1.14",
-    "@types/html-to-text": "~9.0.4",
-    "@types/mjml": "~4.7.4",
-    "@types/nodemailer": "~6.4.17",
-    "@types/vscode": "^1.99.1",
-    "@typescript-eslint/eslint-plugin": "^7.18.0",
-    "@typescript-eslint/parser": "^7.18.0",
-    "autoprefixer": "^10.4.20",
-    "browserslist": "^4.24.3",
-    "eslint": "^8.57.1",
-    "eslint-plugin-import": "^2.31.0",
-    "eslint-plugin-jsx-a11y": "^6.10.2",
-    "inquirer": "^9.3.7",
-    "prettier": "^3.4.2",
-    "typescript": "^5.7.2",
-    "vite": "^5.4.11",
-    "vite-tsconfig-paths": "^4.3.2"
+    "@codingame/monaco-vscode-rollup-extension-directory-plugin": "^26.0.0",
+    "@codingame/monaco-vscode-rollup-vsix-plugin": "^26.0.0",
+    "@effect/language-service": "^0.73.1",
+    "@eslint/js": "^10.0.1",
+    "@types/bun": "^1.3.9",
+    "@types/dockerode": "^3.3.47",
+    "@types/html-to-text": "^9.0.4",
+    "@types/mjml": "^4.7.4",
+    "@types/node": "^24.10.9",
+    "@types/nodemailer": "^7.0.5",
+    "@types/pg": "^8.16.0",
+    "@types/wicg-file-system-access": "^2023.10.7",
+    "@types/ws": "^8.18.1",
+    "@vitejs/plugin-vue": "^6.0.4",
+    "cypress": "^15.10.0",
+    "eslint": "^10.0.0",
+    "eslint-plugin-cypress": "^5.3.0",
+    "sass-embedded": "^1.97.3",
+    "typescript": "~5.9.3",
+    "typescript-eslint": "^8.55.0",
+    "vite": "~7.3.1"
   },
-  "trustedDependencies": [
-    "cpu-features",
-    "esbuild",
-    "postgres",
-    "ssh2"
-  ]
+  "overrides": {
+    "dotenv": "^16.6.1",
+    "esbuild": "^0.24.0"
+  }
 }
