@@ -1,47 +1,46 @@
-// const { awsRegion } = require('@app/config')
-// const aws = require('aws-sdk')
-const fs = require("fs/promises");
-const chalk = require("chalk");
-const nodemailer = require("nodemailer");
-
-const { readFile, writeFile } = fs;
+// @ts-check
+import fs from "fs/promises";
+import * as nodemailer from "nodemailer";
+import { styleText } from "node:util";
 
 const isTest = process.env.NODE_ENV === "test";
 const isDev = process.env.NODE_ENV !== "production";
+const awsRegion = process.env.AWS_REGION || "us-east-2";
 
-let transporterPromise;
+/** @type {nodemailer.Transporter} */
+let transporter;
 const etherealFilename = `${process.cwd()}/.ethereal`;
 
 let logged = false;
 
-module.exports = function getTransport() {
-  if (!transporterPromise) {
-    transporterPromise = (async () => {
-      if (isTest) return nodemailer.createTransport({ jsonTransport: true });
-
+export default async function getTransport() {
+  if (!transporter) {
+    transporter = await (async () => {
+      if (isTest) {
+        return nodemailer.createTransport({ jsonTransport: true });
+      }
       if (isDev) {
         let account;
         try {
-          const testAccountJson = await readFile(etherealFilename, "utf8");
+          const testAccountJson = await fs.readFile(etherealFilename, "utf8");
           account = JSON.parse(testAccountJson);
         } catch (e) {
           account = await nodemailer.createTestAccount();
-          await writeFile(etherealFilename, JSON.stringify(account));
+          await fs.writeFile(etherealFilename, JSON.stringify(account));
         }
         if (!logged) {
           logged = true;
-          console.log();
-          console.log();
+          console.log("\n");
           console.log(
-            chalk.bold(
+            styleText(
+              ["bold"],
               " ✉️ Emails in development are sent via ethereal.email; your credentials follow:",
             ),
           );
           console.log("  Site:     https://ethereal.email/login");
           console.log(`  Username: ${account.user}`);
           console.log(`  Password: ${account.pass}`);
-          console.log();
-          console.log();
+          console.log("\n");
         }
         return nodemailer.createTransport({
           host: "smtp.ethereal.email",
@@ -53,19 +52,20 @@ module.exports = function getTransport() {
           },
         });
       }
-      // if (!process.env.AWS_ACCESS_KEY_ID) {
-      //   throw new Error('Misconfiguration: no AWS_ACCESS_KEY_ID')
-      // }
-      // if (!process.env.AWS_SECRET_ACCESS_KEY) {
-      //   throw new Error('Misconfiguration: no AWS_SECRET_ACCESS_KEY')
-      // }
-      // return nodemailer.createTransport({
-      //   SES: new aws.SES({
-      //     apiVersion: '2010-12-01',
-      //     region: awsRegion,
-      //   }),
-      // })
+      if (!process.env.AWS_ACCESS_KEY_ID) {
+        throw new Error("no AWS_ACCESS_KEY_ID configured");
+      }
+      if (!process.env.AWS_SECRET_ACCESS_KEY) {
+        throw new Error("no AWS_SECRET_ACCESS_KEY configured");
+      }
+      const { SendEmailCommand, SESv2Client } = await import(
+        "@aws-sdk/client-sesv2"
+      );
+      const sesClient = new SESv2Client({ region: awsRegion });
+      return nodemailer.createTransport({
+        SES: { sesClient, SendEmailCommand },
+      });
     })();
   }
-  return transporterPromise;
-};
+  return transporter;
+}
